# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Connect | Server",
    "version": "2.0.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/modules/{id}": {
      "put": {
        "operationId": "updateModule",
        "summary": "Update a module",
        "tags": [
          "Module"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "enum": [
                      "BASE",
                      "CHAT",
                      "LOCATION",
                      "STATUS",
                      "RADIO",
                      "PA_SYSTEM",
                      "DRIVERS",
                      "IRIS"
                    ]
                  },
                  "description": {
                    "type": "string"
                  },
                  "billingType": {
                    "type": "string",
                    "enum": [
                      "FIXED",
                      "PER_AMBULANCE",
                      "PER_BASE",
                      "PER_DRIVER_AMBULANCE"
                    ]
                  },
                  "defaultPrice": {
                    "type": "number"
                  },
                  "internal": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteModule",
        "summary": "Delete a module",
        "tags": [
          "Module"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/units/{unitId}": {
      "put": {
        "operationId": "updateUnit",
        "summary": "Update unit",
        "tags": [
          "Unit"
        ],
        "description": "Update unit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "document": {
                    "description": "Brazilian CNPJ",
                    "type": "string",
                    "minLength": 14,
                    "maxLength": 14
                  }
                },
                "required": [
                  "name",
                  "document"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "unitId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteUnit",
        "summary": "Delete unit",
        "tags": [
          "Unit"
        ],
        "description": "Delete unit",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "unitId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getUnit",
        "summary": "Get unit",
        "tags": [
          "Unit"
        ],
        "description": "Get unit",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "unitId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "cuid",
                          "pattern": "^[cC][^\\s-]{8,}$"
                        },
                        "name": {
                          "type": "string"
                        },
                        "document": {
                          "nullable": true,
                          "type": "string"
                        },
                        "companyId": {
                          "type": "string",
                          "format": "cuid",
                          "pattern": "^[cC][^\\s-]{8,}$"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "deletedAt": {
                          "nullable": true,
                          "type": "string",
                          "format": "date-time"
                        },
                        "phones": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "cuid",
                                "pattern": "^[cC][^\\s-]{8,}$"
                              },
                              "number": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": [
                              "id",
                              "number",
                              "createdAt",
                              "updatedAt"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "document",
                        "companyId",
                        "createdAt",
                        "updatedAt",
                        "deletedAt",
                        "phones"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/units/{unitId}/phones": {
      "post": {
        "operationId": "createUnitPhoneNumber",
        "summary": "Create unit phone number",
        "tags": [
          "Unit"
        ],
        "description": "Create unit phone number",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "number": {
                    "description": "Brazilian phone number (example: +5511999999999)",
                    "type": "string"
                  }
                },
                "required": [
                  "number"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "unitId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "cuid",
                      "pattern": "^[cC][^\\s-]{8,}$"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}